+++ MusicDatabase/MusicBrainz/MusicBrainzSettingsViewController.m
@@ -34,9 +34,9 @@ NSString * const	kMusicBrainzServiceName					= @"MusicBrainz";
 	
 	// Search for the item in the keychain
 	OSStatus status = SecKeychainFindGenericPassword(NULL,
-													 strlen(serviceNameUTF8),
+													 (UInt32)strlen(serviceNameUTF8),
 													 serviceNameUTF8,
-													 strlen(usernameUTF8),
+													 (UInt32)strlen(usernameUTF8),
 													 usernameUTF8,
 													 &passwordLength,
 													 &passwordData,
@@ -78,9 +78,9 @@ NSString * const	kMusicBrainzServiceName					= @"MusicBrainz";
 
 	// Search for the item in the keychain
 	OSStatus status = SecKeychainFindGenericPassword(NULL,
-													 strlen(serviceNameUTF8),
+													 (UInt32)strlen(serviceNameUTF8),
 													 serviceNameUTF8,
-													 strlen(usernameUTF8),
+													 (UInt32)strlen(usernameUTF8),
 													 usernameUTF8,
 													 NULL,
 													 NULL,
@@ -89,11 +89,11 @@ NSString * const	kMusicBrainzServiceName					= @"MusicBrainz";
 	// If the item wasn't found, store it
 	if(errSecItemNotFound == status) {		
 		/*status = */SecKeychainAddGenericPassword(NULL,
-												   strlen(serviceNameUTF8),
+												   (UInt32)strlen(serviceNameUTF8),
 												   serviceNameUTF8,
-												   strlen(usernameUTF8),
+												   (UInt32)strlen(usernameUTF8),
 												   usernameUTF8,
-												   strlen(passwordUTF8), 
+												   (UInt32)strlen(passwordUTF8), 
 												   passwordUTF8, 
 												   &keychainItemRef);
 	}
@@ -101,7 +101,7 @@ NSString * const	kMusicBrainzServiceName					= @"MusicBrainz";
 	else if(noErr == status) {
 		/*status = */SecKeychainItemModifyAttributesAndData(keychainItemRef,
 															NULL,
-															strlen(passwordUTF8), 
+															(UInt32)strlen(passwordUTF8), 
 															passwordUTF8);
 	}
